Fixed snapTo()
snapTo() now works properly. Also, zoom levels are more easily changed,
and zoom has been readjusted to start at one.

Next up, we need to move all the buttons to the subclass itself, and
probably rename the class to something that actually makes more sense,
like RoomEditor or something. Anyways, my arguments for moving the
buttons to RoomEditor are listed in comments on it.